+++ kurtosis-module/impl/participant_network/el/nethermind/nethermind_rpc_api_objects.go
@@ -1,0 +1,9 @@
+package nethermind
+
+type GetNodeInfoResponse struct {
+	Result NodeInfo `json:"result"`
+}
+
+type NodeInfo struct {
+	Enode string	`json:"enode"`
+}