+++ docs/changelog.md
@@ -1,6 +1,7 @@
 # TBD
 ### Features
 * Added a functioning Nimbus CL client
+* Added Nethermind EL client
 * Added a Lodestar CL beacon Launcher
 
 # 0.2.1
@@ -20,6 +21,7 @@
 
 ### Fixes
 * Correct merge parameters like TTD, Altair fork version, merge fork version, etc. per Parithosh's recommendations
+* Give Teku nodes 120s to start
 
 ### Changes
 * Refactor the structure to reflect that there should be one EL node per CL node (and prepare for separated Beacon/validator nodes, like Lighthouse does)